+++ forge/src/main/java/traben/entity_model_features/forge/EMFForge.java
@@ -2,12 +2,14 @@ package traben.entity_model_features.forge;
 
 
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.ConfigScreenHandler;
 import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.loading.FMLEnvironment;
 import net.minecraftforge.network.NetworkConstants;
 import traben.entity_model_features.EMFClient;
+import traben.entity_model_features.config.EMFConfigMainScreen;
 
 @Mod(EMFClient.MOD_ID)
 public class EMFForge {
@@ -17,24 +19,22 @@ public class EMFForge {
 
         if (FMLEnvironment.dist == Dist.CLIENT) {
 
-/*copy of etf config loading code for later
-            try {
-                ModLoadingContext.get().registerExtensionPoint(
-                        ConfigScreenHandler.ConfigScreenFactory.class,
-                        () -> new ConfigScreenHandler.ConfigScreenFactory((minecraftClient, screen) -> new ETFConfigScreenMain(screen)));
-            } catch (NoClassDefFoundError e) {
-                System.out.println("[Entity Texture Features]: Mod config broken, download latest forge version");
-            }
-*/
-
             //not 100% sure what this actually does but it will trigger the catch if loading on the server side
             ModLoadingContext.get().registerExtensionPoint(IExtensionPoint.DisplayTest.class, () -> new IExtensionPoint.DisplayTest(() -> NetworkConstants.IGNORESERVERONLY, (a, b) -> true));
 
 
+            try {
+                ModLoadingContext.get().registerExtensionPoint(
+                        ConfigScreenHandler.ConfigScreenFactory.class,
+                        () -> new ConfigScreenHandler.ConfigScreenFactory((minecraftClient, screen) -> new EMFConfigMainScreen(screen)));
+            } catch (NoClassDefFoundError e) {
+                System.out.println("[Entity Model Features]: Mod config screen broken, download latest forge version");
+            }
+
             EMFClient.init();
         } else {
 
-            throw new UnsupportedOperationException("Attempting to load a clientside only mod on the server, refusing");
+            throw new UnsupportedOperationException("Attempting to load a clientside only mod [EMF] on the server, refusing");
         }
     }
 }